pull

+++ MovieBooking/Services/ReservationsAdministration/Administration.Domain/Entities/TicketReservation.cs
@@ -15,11 +15,11 @@ namespace Administration.Domain.Entities
         public string MovieTitle { get; set; }
         public string TheaterHallId { get; set; }
         public string TheaterHallName { get; set; }
-        public decimal Price { get; set; }
+        public int Price { get; set; }
         public int NumberOfTickets { get; set; } = 0;
 
 
-        public TicketReservation(string projectionId, string movieTitle, string movieId, decimal price, string theaterHallId, string theaterHallName, int numberOfTickets)
+        public TicketReservation(string projectionId, string movieTitle, string movieId, int price, string theaterHallId, string theaterHallName, int numberOfTickets)
         {
             ProjectionId = projectionId ?? throw new ArgumentNullException(nameof(projectionId));
             MovieTitle = movieTitle ?? throw new ArgumentNullException(nameof(movieTitle));
